Change onboarding popup click behaviour
Add a separate click handler to ExtractTextWidget so we can do
different things for both the OCR and config buttons when they're
clicked and the onboarding needs to show first. To do this, a
'next action' is added to the OnboardingPopup, which will be called
when the 'Okay, got it' button is clicked.

Behaviour when the onboarding popup is not in use does not change.

Bug: T288575
Change-Id: I56c3c70ce5551e8a082e2ea7acfa3d6c4f8f8c2a