SiteLinks: Limits the number of fetched entities
Uses the built-in Wikibase limited entity fetcher.
This way the limit is the global one
and if an entity is fetched for sitelinks
it is fine to fetch it again during the page rendering.

Currently, the code ignores the entities when the limit is encountered. I am not sure what would be the proper behavior.

Not very useful now but required to make I28c09e7cb95ba47a48b2e2077157a099f8c32383 safe.

Change-Id: I70497b61ba8c45bc322d9818e735f65aaa69f052