Use service 'GadgetsRepo' instead of deprecated GadgetRepo::singleton()
The service 'GadgetsRepo' gets injected as optional service.

The availability of the service 'GadgetsRepo' is equivalent to the
	ExtensionRegistry::getInstance()->isLoaded( 'Gadgets' )
and therefore this check is replaced.

This change requires a phan dependency on extension Gadgets in
project integration/config in file zuul/parameter_functions.py:
I1f526abf32155e09bb4099ac96dba93750cc709a

Change-Id: Iae5cf4c0a0d8fdac7e291fdeb0e7f4e4a78a1d58